requirements.txt
aiohttp
bs4
lxml
PyYaml